Create README.md
This is basically a browser I made in an hour without any chromium or stuff. Expect bugs because I made it in an hour but I don't know if any are there.

This browser does not contain any google stuff so yeah, though I might be wrong because they own far too much. This browser is also pretty private for most people, and I am fine if you want to customize it just, please give me credit if you are going to upload it.

Dependencies if you are going to use the python file:
pip install PyQt5 PyQtWebEngine